requirements.txt
requests==2.32.3
numpy==1.25.2
python-dotenv==1.0.1
sentence-transformers==3.4.1
faiss-cpu==1.9.0.post1
pydantic==2.10.6
httpx==0.27.2
gradio==5.23.1
langgraph==0.2.62
langchain_groq==0.2.4
langchain_core==0.3.47